+++ CHANGELOG.md
@@ -37,13 +37,15 @@
   `params` posted in the request header.
 
 * `DefaultRoleService` has improved error handling for failed directory group lookups.
-* `LdapService` now optionally throws if a query fails rather than returning an empty result. This is not
-  expected to affect most applications, but may require some to adjust their error handling.
-
+* `LdapService` bulk lookup methods now provide a `strict` option to throw if a query fails rather than
+   quietly returning an empty result.
 
 ### 💥 Breaking Changes
 
 * Requires `hoist-react >= 63.0` for client-side support of the new `track` and `submitError` endpoints.
+* Implementations of `DefaultRoleService.doLoadUsersForDirectoryGroups`  will need to handle a new
+ `strictMode` flag provided as a second argument.
+
 
 
 ## 18.5.1 - 2024-03-08